src/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
dnspython==2.7.0
dotenv==0.9.9
fastapi==0.109.2
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.5.0
gradio==4.41.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
motor==3.3.2
multidict==6.1.0
multiprocess==0.70.16
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pymongo==4.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.11.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.11.2
semantic-version==2.10.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.27.1
websockets==12.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
locust==2.21.0